+++ krokiet/src/main.rs
@@ -28,8 +28,7 @@ use crate::connect_scan::connect_scan_button;
 use crate::connect_select::{connect_select, connect_showing_proper_select_buttons};
 use crate::connect_show_preview::connect_show_preview;
 use crate::connect_size_of_config_cache::connect_size_of_config_cache;
-use crate::connect_sort::{connect_showing_proper_sort_buttons, connect_sort};
-use crate::connect_sort_column::connect_sort_column;
+use crate::connect_sort::{connect_showing_proper_sort_buttons, connect_sort, connect_sort_column};
 use crate::connect_stop::connect_stop_button;
 use crate::connect_translation::connect_translations;
 // TODO - at start this should be used, to be sure that rust models are in sync with slint models
@@ -52,7 +51,6 @@ mod connect_select;
 mod connect_show_preview;
 mod connect_size_of_config_cache;
 mod connect_sort;
-mod connect_sort_column;
 mod connect_stop;
 mod connect_translation;
 mod localizer_krokiet;